Dodal nivo abstrakcije za Grid
